+++ test/js/return.test.js
@@ -14,4 +14,25 @@ describe("return", () => {
 
   test("return with breaking", () =>
     expect(`return ${long}`).toChangeFormat(`return(\n  ${long}\n)`));
+
+  test("returning an array", () =>
+    expect("return [1, 2, 3]").toChangeFormat("return 1, 2, 3"));
+
+  test("returning a list that breaks", () =>
+    expect(`return ${long}, ${long}`).toChangeFormat(
+      `return [\n  ${long},\n  ${long}\n]`
+    ));
+
+  test("returning an array within parens", () =>
+    expect("return([1, 2, 3])").toChangeFormat("return 1, 2, 3"));
+
+  test("returning a long special array", () =>
+    expect(`return %w[${long}]`).toChangeFormat(
+      `return(\n  %w[\n    ${long}\n  ]\n)`
+    ));
+
+  test("returning two arguments, one that breaks", () =>
+    expect(`return foo, ${long}`).toChangeFormat(
+      `return [\n  foo,\n  ${long}\n]`
+    ));
 });